+++ merged-packages/smart-transactions-controller/src/SmartTransactionsController.test.ts
@@ -1,13 +1,17 @@
-import { deriveStateFromMetadata, Messenger } from '@metamask/base-controller';
+import { deriveStateFromMetadata } from '@metamask/base-controller';
 import {
   NetworkType,
   convertHexToDecimal,
   ChainId,
 } from '@metamask/controller-utils';
 import {
-  type NetworkControllerGetNetworkClientByIdAction,
-  type NetworkControllerGetStateAction,
-  type NetworkControllerStateChangeEvent,
+  Messenger,
+  MOCK_ANY_NAMESPACE,
+  type MessengerActions,
+  type MessengerEvents,
+  type MockAnyNamespace,
+} from '@metamask/messenger';
+import {
   NetworkStatus,
   RpcEndpointType,
   type NetworkState,
@@ -37,14 +41,17 @@ import {
   DEFAULT_INTERVAL,
   SmartTransactionsController,
   getDefaultSmartTransactionsControllerState,
-} from './SmartTransactionsController';
-import type {
-  SmartTransactionsControllerActions,
-  SmartTransactionsControllerEvents,
+  type SmartTransactionsControllerMessenger,
 } from './SmartTransactionsController';
 import type { SmartTransaction, UnsignedTransaction, Hex } from './types';
 import { SmartTransactionStatuses, ClientId } from './types';
 import * as utils from './utils';
+
+type AllActions = MessengerActions<SmartTransactionsControllerMessenger>;
+
+type AllEvents = MessengerEvents<SmartTransactionsControllerMessenger>;
+
+type RootMessenger = Messenger<MockAnyNamespace, AllActions, AllEvents>;
 
 jest.mock('@metamask/eth-query', () => {
   const EthQuery = jest.requireActual('@metamask/eth-query');
@@ -2540,7 +2547,7 @@ describe('SmartTransactionsController', () => {
           deriveStateFromMetadata(
             controller.state,
             controller.metadata,
-            'anonymous',
+            'includeInDebugSnapshot',
           ),
         ).toMatchInlineSnapshot(`
           {
@@ -2717,17 +2724,10 @@ async function withController<ReturnValue>(
     updateTransaction = jest.fn(),
   } = rest;
 
-  const controllerMessenger = new Messenger<
-    | SmartTransactionsControllerActions
-    | NetworkControllerGetNetworkClientByIdAction
-    | NetworkControllerGetStateAction
-    | TransactionControllerGetNonceLockAction
-    | TransactionControllerConfirmExternalTransactionAction
-    | TransactionControllerGetTransactionsAction
-    | TransactionControllerUpdateTransactionAction,
-    SmartTransactionsControllerEvents | NetworkControllerStateChangeEvent
-  >();
-  controllerMessenger.registerActionHandler(
+  const rootMessenger: RootMessenger = new Messenger({
+    namespace: MOCK_ANY_NAMESPACE,
+  });
+  rootMessenger.registerActionHandler(
     'NetworkController:getNetworkClientById',
     jest.fn().mockImplementation((networkClientId) => {
       switch (networkClientId) {
@@ -2750,7 +2750,7 @@ async function withController<ReturnValue>(
       }
     }),
   );
-  controllerMessenger.registerActionHandler(
+  rootMessenger.registerActionHandler(
     'NetworkController:getState',
     jest.fn().mockReturnValue({
       selectedNetworkClientId: NetworkType.mainnet,
@@ -2788,26 +2788,35 @@ async function withController<ReturnValue>(
       },
     }),
   );
-  controllerMessenger.registerActionHandler(
+  rootMessenger.registerActionHandler(
     'TransactionController:getNonceLock',
     getNonceLock,
   );
-  controllerMessenger.registerActionHandler(
+  rootMessenger.registerActionHandler(
     'TransactionController:confirmExternalTransaction',
     confirmExternalTransaction,
   );
-  controllerMessenger.registerActionHandler(
+  rootMessenger.registerActionHandler(
     'TransactionController:getTransactions',
     getTransactions,
   );
-  controllerMessenger.registerActionHandler(
+  rootMessenger.registerActionHandler(
     'TransactionController:updateTransaction',
     updateTransaction,
   );
 
-  const messenger = controllerMessenger.getRestricted({
-    name: 'SmartTransactionsController',
-    allowedActions: [
+  const messenger = new Messenger<
+    'SmartTransactionsController',
+    AllActions,
+    AllEvents,
+    RootMessenger
+  >({
+    namespace: 'SmartTransactionsController',
+    parent: rootMessenger,
+  });
+  rootMessenger.delegate({
+    messenger,
+    actions: [
       'NetworkController:getNetworkClientById',
       'NetworkController:getState',
       'TransactionController:getNonceLock',
@@ -2815,7 +2824,7 @@ async function withController<ReturnValue>(
       'TransactionController:getTransactions',
       'TransactionController:updateTransaction',
     ],
-    allowedEvents: ['NetworkController:stateChange'],
+    events: ['NetworkController:stateChange'],
   });
 
   const controller = new SmartTransactionsController({
@@ -2834,7 +2843,7 @@ async function withController<ReturnValue>(
   });
 
   function triggerNetworStateChange(state: NetworkState) {
-    controllerMessenger.publish('NetworkController:stateChange', state, []);
+    rootMessenger.publish('NetworkController:stateChange', state, []);
   }
 
   triggerNetworStateChange({